cron-runner/pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cron"
version = "0.1.0"
description = "Cron sheduler for python cron jobs"

dependencies = [
   'openmeteo-requests>=1.3.0',
   'requests-cache>=1.2.1',
   'retry-requests>=2.0.0',
   'numpy >= 2.1.2',
   'pandas >= 2.2.3',
   'pathlib>=1.0.1',
   'influxdb-client',
   'pytz',
   'discord.py'
]

[project.scripts]
cron-main = "bin.main:main"
transfer-csv-to-influx = "bin.transfer_csv_to_influx:main"
fix-time = "bin.fix_time:main"
get-models-with-ids = "bin.get_models_with_ids:main"
get-model-with-no-data-for-location = "bin.get_model_with_no_data_for_location:main"

[tool.setuptools]
packages = ["cron", "bin"]